chapter 2/building_scalable_datalake/retail-data-lake-python/requirements.txt
aws-cdk-lib==2.164.1
constructs>=10.0.0,<11.0.0
boto3
pandas
pyarrow
